requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
Pillow==6.2.1
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
PyMySQL==0.9.3
pyparsing==2.4.5
pytest==5.2.2
PyYAML==5.1.2
six==1.13.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
